# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8090
info:
  contact:
    name: PopuGate Support
    url: https://github.com/fussraider/PopuGate
  description: PopuGate is a Telegram MTProto proxy manager with a REST API.
  license:
    name: MIT
    url: https://github.com/fussraider/PopuGate/blob/main/LICENSE
  title: PopuGate API
  version: "1.0"
paths:
...
  /backups:
    get:
      description: Returns a list of all available backups and whether backup encryption
        is enabled.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List backups
      tags:
      - backup
    post:
      description: Creates a new encrypted or unencrypted backup depending on BACKUP_ENCRYPTION_KEY
        env var. Returns filename, size, checksum, and creation timestamp.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create backup
      tags:
      - backup
  /backups/{filename}:
    delete:
      description: Deletes a backup file and its checksum sidecar by filename
      parameters:
      - description: Backup filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete backup
      tags:
      - backup
  /backups/download/{filename}:
    get:
      description: Downloads a backup file (encrypted or plain) by filename as a binary
        attachment
      parameters:
      - description: Backup filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download backup
      tags:
      - backup
  /backups/restore:
    post:
      consumes:
      - application/json
      description: Stops the proxy engine, restores database and config from backup,
        rotates JWT secret, then restarts the engine. Verifies SHA256 checksum if
        available.
      parameters:
      - description: Backup filename to restore
        in: body
        name: body
        required: true
        schema:
          properties:
            filename:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Restore backup
      tags:
      - backup
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"